src/main/java/models/User.java
package models;

import java.util.Objects;

public class User {
    private int id;
    private String userName;
    private String position;
    private int departmentId;

    public User( String userName, String position, int departmentId) {
        this.userName = userName;
        this.position = position;
        this.departmentId = departmentId;

    }

    public String getUserName() {
        return userName;
    }

    public void setUserName(String userName) {
        this.userName = userName;
    }

    public void setId(int id) {
        this.id = id;
    }


    public void setName(String name) {
        this.userName = name;
    }

    public String getPosition() {
        return position;
    }

    public void setPosition(String position) {
        this.position = position;
    }

    public int getDepartmentId() {
        return departmentId;
    }

    public void setDepartmentId(int departmentId) {
        this.departmentId = departmentId;
    }



    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        User user = (User) o;
        return id == user.id &&
                departmentId == user.departmentId &&
                Objects.equals(userName, userName) &&
                Objects.equals(position, user.position);
    }

    @Override
    public int hashCode() {
        return Objects.hash( userName, position, departmentId);
    }

}